+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLListSelectionField.cs
@@ -91,7 +91,11 @@ namespace EntityGraphQL.Compiler
                 bool needsServiceWrap = NeedsServiceWrap(withoutServiceFields);
                 if (needsServiceWrap)
                 {
-                    resultExpression = ExpressionUtil.MakeSelectWithDynamicType(this, nextFieldContext!, listContext, selectionFields, true);
+                    // To support a common use case where we are coming from a service result to another service field where the 
+                    // service is the Query Context. Which we are assuming is likely an EF context and we don't need the null check
+                    // Use ExecutionOptions.ExecuteServiceFieldsSeparately = false to disable this behaviour
+                    var nullCheck = Field!.Services.Any(s => s.Type != Field.Schema.QueryContextType);
+                    resultExpression = ExpressionUtil.MakeSelectWithDynamicType(this, nextFieldContext!, listContext, selectionFields, nullCheck);
                 }
             }