+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLInlineFragmentField.cs
@@ -21,7 +21,7 @@ namespace EntityGraphQL.Compiler
             return QueryFields;
         }
 
-        private void GetServices(CompileContext compileContext, BaseGraphQLField gqlField)
+        private static void GetServices(CompileContext compileContext, BaseGraphQLField gqlField)
         {
             if (gqlField.Field != null && gqlField.Field.Services.Any())
             {